Move release functions and structs to its own file
